requirements.txt
feedparser==6.0.10
tk-tools==0.16.0
python-qbittorrent==0.4.2
inquirer==2.10.0
attr==0.3.2
brotli==1.0.9
brotlicffi==1.0.9.2
cchardet==2.1.7
ConfigParser==5.2.0
cryptography==37.0.4
Cython==0.29.32
dl==0.1.0
docutils==0.19
HTMLParser==0.0.2
ipython==8.4.0
ipywidgets==7.7.1
Jinja2==3.1.2
jnius==1.1.0
keyring==23.8.2
lockfile==0.12.2
mock==4.0.3
numpy==1.23.1
Pillow==9.2.0
protobuf==4.21.5
pyOpenSSL==22.0.0
railroad==0.5.0
simplejson==3.17.6
Sphinx==5.1.1
testsuite==0.1.2
tornado==6.2
trove_classifiers==2022.8.7
truststore==0.4.0
unicodedata2==14.0.0
xmlrpclib==1.0.1
